+++ console2/src/components/organisms/index.ts
@@ -21,6 +21,7 @@ export { default as APITokenList } from './APITokenList';
 export { default as APITokenDeleteActivity } from './APITokenDeleteActivity';
 export { default as CancelProcessPopup } from './CancelProcessPopup';
 export { default as DeleteRepositoryPopup } from './DeleteRepositoryPopup';
+export { default as EditProjectActivity } from './EditProjectActivity';
 export { default as EditRepositoryActivity } from './EditRepositoryActivity';
 export { default as EncryptValueActivity } from './EncryptValueActivity';
 export { default as FindUserField } from './FindUserField';